# qiansi-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "纸喵软件系列之服务端",
        "title": "纸喵 qiansi API",
        "termsOfService": "http://zhimiao.org",
        "contact": {
            "name": "API Support",
            "url": "http://tools.zhimiao.org",
            "email": "mail@xiaoliu.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1315",
    "basePath": "/",
    "paths": {
        "/admin/DeployDel": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除部署应用",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeployDelParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"操作成功\",\"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/admin/DeployDo": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "启动部署 TODO: 后期关闭此接口的开放特性，新增外部接口，通过不可枚举key作为部署参数",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeployDoParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"启动成功\",\"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
...
        "/admin/DeployRelationServer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "部署应用关联服务器",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeployRelationParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"关联成功\",\"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/admin/DeployRunLog": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取当前部署应用指定服务器的运行日志",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeployRunLogParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.DeployLog"
                        }
                    }
                }
            }
        },
        "/admin/DeployRunLogTab": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取当前部署应用绑定的服务器列表，用于渲染运行日志选项卡",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeployServerParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Server"
                        }
                    }
                }
            }
        },
        "/admin/DeployServer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取当前部署应用的服务器列表",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeployServerParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Server"
                        }
                    }
                }
            }
        },
        "/admin/DeploySet": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "设置部署应用",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.DeploySetParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"操作成功\",\"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/admin/ScheduleLists": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取计划任务列表",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.PageParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.PageInfo"
                        }
                    }
                }
            }
        },
        "/admin/ServerDel": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除服务器",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.ServerDelParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"操作成功\",\"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
...
        "/admin/ServerSet": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "设置服务器信息",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/req.ServerSetParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"操作成功\",\"data\": null}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.DeployLog": {
            "type": "object",
            "properties": {
                "clientIp": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "deployId": {
                    "type": "integer"
                },
                "deployVersion": {
                    "type": "integer"
                },
                "deviceId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "serverId": {
                    "type": "integer"
                },
                "uid": {
                    "type": "integer"
                }
            }
        },
        "models.Server": {
            "type": "object",
            "properties": {
                "apiSecret": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "deviceId": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "serverName": {
                    "type": "string"
                },
                "serverRuleId": {
                    "type": "integer"
                },
                "serverStatus": {
                    "type": "integer"
                },
                "uid": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "req.DeployDelParam": {
            "type": "object",
            "properties": {
                "deployId": {
                    "description": "部署应用ID",
                    "type": "integer"
                }
            }
        },
        "req.DeployDoParam": {
            "type": "object",
            "properties": {
                "deployId": {
                    "description": "部署应用ID",
                    "type": "integer"
                }
            }
        },
        "req.DeployRelationParam": {
            "type": "object",
            "properties": {
                "deployId": {
                    "description": "部署应用ID",
                    "type": "integer"
                },
                "relation": {
                    "description": "关联操作 true-关联 false-取消",
                    "type": "boolean"
                },
                "serverId": {
                    "description": "服务器ID",
                    "type": "integer"
                }
            }
        },
        "req.DeployRunLogParam": {
            "type": "object",
            "properties": {
                "deployId": {
                    "description": "部署应用ID",
                    "type": "integer"
                },
                "serverId": {
                    "description": "服务器ID",
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "req.DeployServerParam": {
            "type": "object",
            "properties": {
                "deployId": {
                    "description": "部署应用ID",
                    "type": "integer"
                }
            }
        },
        "req.DeploySetParam": {
            "type": "object",
            "properties": {
                "afterCommand": {
                    "description": "后置命令",
                    "type": "string"
                },
                "beforeCommand": {
                    "description": "前置命令",
                    "type": "string"
                },
                "branch": {
                    "description": "git分支",
                    "type": "string"
                },
                "deployKeys": {
                    "description": "部署私钥",
                    "type": "string"
                },
                "deployType": {
                    "description": "部署类型 0-本地 1-git 2-zip",
                    "type": "integer"
                },
                "id": {
                    "description": "应用唯一编号",
                    "type": "integer"
                },
                "localPath": {
                    "description": "本地部署地址",
                    "type": "string"
                },
                "remoteUrl": {
                    "description": "资源地址",
                    "type": "string"
                },
                "title": {
                    "description": "应用名称",
                    "type": "string"
                }
            }
        },
        "req.PageParam": {
            "type": "object",
            "properties": {
                "lastId": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "req.ServerDelParam": {
            "type": "object",
            "properties": {
                "serverId": {
                    "type": "integer"
                }
            }
        },
        "req.ServerSetParam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "serverName": {
                    "description": "服务器备注名",
                    "type": "string"
                },
                "serverRuleId": {
                    "description": "服务器规则id",
                    "type": "integer"
                },
                "serverStatus": {
                    "description": "服务器状态 -1-失效 0-待认领 1-已分配通信密钥 2-已绑定",
                    "type": "integer"
                }
            }
        },
...
        "resp.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "resp.PageInfo": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "rows": {
                    "type": "object"
                },
                "totalSize": {
                    "type": "integer"
                }
            }
        },
...
    }
}